fix error

+++ src/main/java/com/gosoft/ecommerce/controller/UserController.java
@@ -1,8 +1,11 @@
 package com.gosoft.ecommerce.controller;
 
 import com.gosoft.ecommerce.entity.User;
+import com.gosoft.ecommerce.enums.Type;
 import com.gosoft.ecommerce.model.UserUpdateRequest;
+import com.gosoft.ecommerce.response.GlobalResponse;
 import com.gosoft.ecommerce.service.UserService;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -16,8 +19,8 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
+import java.util.Optional;
 
-@RequestMapping("/users")
 @RestController
 public class UserController {
     private final UserService userService;
@@ -26,31 +29,53 @@ public class UserController {
         this.userService = userService;
     }
 
-    @GetMapping("/me")
-    public ResponseEntity<User> authenticatedUser() {
+    @GetMapping("/adminuser/profile")
+    public ResponseEntity<GlobalResponse<?>> authenticatedUser() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
         User currentUser = (User) authentication.getPrincipal();
 
-        return ResponseEntity.ok(currentUser);
+        GlobalResponse<User> response = new GlobalResponse<>(currentUser);
+        return ResponseEntity.ok(response);
     }
 
-    @GetMapping()
-    public ResponseEntity<List<User>> getUsers(@RequestParam String search) {
+    @GetMapping("/admin/user/{userID}")
+    public ResponseEntity<?> getUser(@PathVariable Integer userID) {
+        Optional<User> users = userService.getUserByID(userID);
+        if (users.isEmpty()) {
+            return new ResponseEntity<>(new GlobalResponse<>(Type.USER_NOT_FOUND, userID), HttpStatus.BAD_REQUEST);
+        }
+
+        GlobalResponse<User> response = new GlobalResponse<>(users.get());
+        return ResponseEntity.ok(response);
+    }
+
+    @GetMapping("/admin/users")
+    public ResponseEntity<GlobalResponse<?>> getUsers(@RequestParam String search) {
         List<User> users = userService.searchUsers(search);
-
-        return ResponseEntity.ok(users);
+        GlobalResponse<List<User>> response = new GlobalResponse<>(users);
+        return ResponseEntity.ok(response);
     }
 
-    @PutMapping("/{userID}")
-    public ResponseEntity<User> updateUser(@PathVariable Integer userID, @RequestBody UserUpdateRequest request) throws Exception {
+    @PutMapping("/admin/user/{userID}")
+    public ResponseEntity<GlobalResponse<?>> updateUser(@PathVariable Integer userID, @RequestBody UserUpdateRequest request) throws Exception {
+        Optional<User> users = userService.getUserByID(userID);
+        if (users.isEmpty()) {
+            return new ResponseEntity<>(new GlobalResponse<>(Type.USER_NOT_FOUND, userID), HttpStatus.BAD_REQUEST);
+        }
+
         User user = userService.updateUser(userID, request);
-        return ResponseEntity.ok(user);
+        GlobalResponse<User> response = new GlobalResponse<>(user);
+        return ResponseEntity.ok(response);
     }
 
-    @DeleteMapping("/{userID}")
-    public ResponseEntity<String> deleteUser(@PathVariable Integer userID) throws Exception {
+    @DeleteMapping("/admin/user/{userID}")
+    public ResponseEntity<GlobalResponse<?>> deleteUser(@PathVariable Integer userID) throws Exception {
+        Optional<User> users = userService.getUserByID(userID);
+        if (users.isEmpty()) {
+            return new ResponseEntity<>(new GlobalResponse<>(Type.USER_NOT_FOUND, userID), HttpStatus.BAD_REQUEST);
+        }
         userService.deleteUser(userID);
-        return ResponseEntity.ok("delete userID "+userID+" success");
+        return ResponseEntity.ok(new GlobalResponse<>());
     }
 }
